+++ Utils/TextUtils.h
@@ -5,5 +5,5 @@
 class TextUtils
 {
 public:
-	static sf::Text *CreateText(sf::Font *font, std::string text, float posX, float posY, float size, sf::Color color);
+	static sf::Text *CreateText(sf::Font *font, std::string text, float posX, float posY, unsigned int size, sf::Color color);
 };
